did registration test
